WIP split acs-table-rows into change and current
