--- app/templates/components/acs-table-header.hbs
@@ -1,101 +1,0 @@
-<table class="header-table">
-  <thead>
-    {{#if (eq mode 'current')}} {{! current header}}
-      <tr>
-        <th rowspan="2">&nbsp;</th>
-        <th colspan="{{if reliability "5" "2"}}" class="text-center cell-border-bottom">Selected Area</th>
-        {{#if comparison}}
-          <th colspan="{{if reliability "5" "2"}}" class="text-center cell-border-bottom">Comparison Area</th>
-          <th colspan="2" class="text-center cell-border-bottom">Difference</th>
-        {{/if}}
-      </tr>
-      <tr>
-        <th class="text-right cell-border-bottom">Estimate</th>
-        {{#if reliability}}
-          <th class="text-right cell-border-bottom">MOE</th>
-          <th class="text-right cell-border-bottom">CV</th>
-        {{/if}}
-        <th class="text-right cell-border-bottom">Percent Estimate</th>
-        {{#if reliability}}
-          <th class="text-right cell-border-bottom">MOE</th>
-        {{/if}}
-        {{#if comparison}}
-          <th class="text-right cell-border-bottom">Estimate</th>
-          {{#if reliability}}
-            <th class="text-right cell-border-bottom">MOE</th>
-            <th class="text-right cell-border-bottom">CV</th>
-          {{/if}}
-          <th class="text-right cell-border-bottom">Percent Estimate</th>
-          {{#if reliability}}
-            <th class="text-right cell-border-bottom">MOE</th>
-          {{/if}}
-          <th class="text-right cell-border-bottom">Number</th>
-          <th class="text-right cell-border-bottom">Percentage Point</th>
-        {{/if}}
-      </tr>
-    {{else}} {{! change over time header}}
-      <tr>
-        <th rowspan="3">&nbsp;</th>
-        {{#unless comparison}}
-          <th colspan="2" rowspan="2" class="text-center cell-border-bottom">2006-2010</th>
-          <th colspan="2" rowspan="2" class="text-center cell-border-bottom">2012-2016</th>
-        {{/unless}}
-        {{#if comparison}}
-          {{#if reliability}}
-            <th colspan="12" class="text-center cell-border-bottom">Change, 2006-2010 to 2012-2016</th>
-          {{else}}
-            <th colspan="6" class="text-center cell-border-bottom">Change, 2006-2010 to 2012-2016</th>
-          {{/if}}
-        {{else}}
-          {{#if reliability}}
-            <th colspan="4" class="text-center cell-border-bottom">Change, 2006-2010 to 2012-2016</th>
-          {{else}}
-            <th colspan="2" class="text-center cell-border-bottom">Change, 2006-2010 to 2012-2016</th>
-          {{/if}}
-        {{/if}}
-      </tr>
-      <tr>
-        <th colspan="{{if reliability "4" "2"}}" class="text-center cell-border-bottom">Selected Area</th>
-        {{#if comparison}}
-          <th colspan="{{if reliability "4" "2"}}" class="text-center cell-border-bottom">Comparison Area</th>
-          <th colspan="{{if reliability "4" "2"}}" class="text-center cell-border-bottom">Difference</th>
-        {{/if}}
-      </tr>
-      <tr>
-        {{#unless comparison}}
-          <th>Number</th>
-          <th>Percent</th>
-          <th>Number</th>
-          <th>Percent</th>
-        {{/unless}}
-        <th>Estimate</th>
-        {{#if reliability}}
-          <th>MOE</th>
-        {{/if}}
-        <th>Percent Estimate</th>
-        {{#if reliability}}
-          <th>MOE</th>
-        {{/if}}
-        {{#if comparison}}
-          <th>Estimate</th>
-          {{#if reliability}}
-            <th>MOE</th>
-          {{/if}}
-          <th>Percent Estimate</th>
-          {{#if reliability}}
-            <th>MOE</th>
-          {{/if}}
-          <th>Estimate</th>
-          {{#if reliability}}
-            <th>MOE</th>
-          {{/if}}
-          <th>Percentage Estimate</th>
-          {{#if reliability}}
-            <th>MOE</th>
-          {{/if}}
-        {{/if}}
-      </tr>
-    {{/if}}
-  </thead>
-</table>
-{{yield}}
