+++ app/templates/components/acs-table-row-change.hbs
@@ -1,0 +1,105 @@
+{{#if rowconfig.divider}}
+
+  <td>&nbsp;</td>
+  {{#unless comparison}}
+    <td class="cell-border-left">&nbsp;</td>
+    <td>&nbsp;</td>
+    <td class="cell-border-left">&nbsp;</td>
+    <td>&nbsp;</td>
+  {{/unless}}
+  <td class="cell-border-left">&nbsp;</td>
+  {{#if reliability}}
+    <td>&nbsp;</td>
+  {{/if}}
+  <td>&nbsp;</td>
+  {{#if reliability}}
+    <td>&nbsp;</td>
+  {{/if}}
+  {{#if comparison}}
+    <td class="cell-border-left">&nbsp;</td>
+    {{#if reliability}}
+      <td>&nbsp;</td>
+    {{/if}}
+    <td>&nbsp;</td>
+    {{#if reliability}}
+      <td>&nbsp;</td>
+    {{/if}}
+    <td class="cell-border-left">&nbsp;</td>
+    {{#if reliability}}
+      <td>&nbsp;</td>
+    {{/if}}
+    <td>&nbsp;</td>
+    {{#if reliability}}
+      <td>&nbsp;</td>
+    {{/if}}
+  {{/if}}
+
+{{else}}{{! subheader && divider = false }}
+
+  <td>{{rowconfig.title}}</td>
+    {{#unless comparison}}
+      <td class="cell-border-left"><!--previous dataset sum-->
+      {{selectedEarlySum}}
+    </td>
+    <td><!--previous dataset percent-->
+      {{selectedEarlyPercent}}
+    </td>
+    <td class="cell-border-left"><!--current dataset sum-->
+      {{selectedCurrentSum}}
+    </td>
+    <td><!--current dataset percent-->
+    {{selectedCurrentPercent}}
+    </td>
+  {{/unless}}
+  <td class="cell-border-left"><!--selected area sum change-->
+    {{differenceChangeSum}}
+  </td>
+  {{#if reliability}}
+  <td><!--selected area sum change moe-->
+    {{differenceChangeSumMoE}}
+  </td>
+  {{/if}}
+  <td><!--selected area percent change-->
+    {{differenceChangePercent}}
+  </td>
+  {{#if reliability}}
+  <td><!--selected area percent change moe-->
+    {{differenceChangePercentMOE}}
+  </td>
+  {{/if}}
+  {{#if comparison}}
+  <td class="cell-border-left"><!--comparison area sum change-->
+    {{comparisonDifferenceChangeSum}}
+  </td>
+  {{#if reliability}}
+  <td><!--comparison area sum change moe-->
+      {{comparisonDifferenceChangeSumMoE}}
+  </td>
+  {{/if}}
+  <td><!--comparison area percent change-->
+      {{comparisonDifferenceChangePercent}}
+  </td><!--comparison area percent change moe-->
+  {{#if reliability}}
+    <td>
+        {{comparisonDifferenceChangePercentMOE}}
+    </td>
+    {{/if}}
+    <td class="cell-border-left">
+      {{longTermDifference}}
+    </td>
+    {{#if reliability}}
+      <td>
+        {{longTermDifferenceMOE}}
+      </td>
+    {{/if}}
+    <td>
+      {{longTermDifferencePercent}}
+    </td>
+    {{#if reliability}}
+    <td>
+      {{longTermDifferencePercentMOE}}
+    </td>
+    {{/if}}
+  {{/if}}
+{{/if}}
+{{yield}}
