--- app/templates/components/acs-table-row.hbs
@@ -1,184 +1,0 @@
-{{#if (eq mode 'current')}} {{! current mode }}
-  {{#if rowconfig.divider}}
-    <td>&nbsp;</td>
-    <td class="cell-border-left">&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-      <td>&nbsp;</td>
-    {{/if}}
-    <td>&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-    {{/if}}
-    {{#if comparison}}
-      <td class="cell-border-left">&nbsp;</td>
-      {{#if reliability}}
-        <td>&nbsp;</td>
-        <td>&nbsp;</td>
-      {{/if}}
-      <td>&nbsp;</td>
-      {{#if reliability}}
-        <td>&nbsp;</td>
-      {{/if}}
-      <td class="cell-border-left">&nbsp;</td>
-      <td>&nbsp;</td>
-    {{/if}}
-  {{else}}
-    <td>{{rowconfig.title}}</td>
-    <td class="cell-border-left {{if (eq data.sum 0) 'medium-gray'}} {{if (gte data.cv 20) 'medium-gray'}}">{{selectedSum}}</td>
-    {{#if reliability}}
-      <td class="{{if (gte data.cv 20) 'medium-gray'}}">
-      {{selectedSumMoE}}
-      </td>
-      <td class="{{if (gte data.cv 20) 'medium-gray'}}">
-      {{selectedCV}}
-      </td>
-    {{/if}}
-    <td class="{{if (gte data.cv 20) 'medium-gray'}}">
-      {{selectedPercent}}
-    </td>
-   {{#if reliability}}
-        <td class="{{if (gte data.cv 20) 'medium-gray'}}">
-          {{selectedPercentM}}
-        </td>
-    {{/if}}
-    {{#if comparison}}
-      <td class="cell-border-left {{if (eq data.comparison_sum 0) 'medium-gray'}} {{if (gte data.comparison_cv 20) 'medium-gray'}}">{{comparisonSum}}</td>
-      {{#if reliability}}
-        <td class="{{if (gte data.comparison_cv 20) 'medium-gray'}}">
-          {{comparisonSumMoE}}
-        </td>
-        <td class="{{if (gte data.comparison_cv 20) 'medium-gray'}}">
-          {{comparisonCV}}
-        </td>
-      {{/if}}
-      <td class="{{if (gte data.comparison_cv 20) 'medium-gray'}}">
-        {{comparisonPercent}}
-      </td>
-      {{#if reliability}}
-          <td class="{{if (gte data.comparison_cv 20) 'medium-gray'}}">
-            {{comparisonPercentM}}
-          </td>
-      {{/if}}
-      <td class="cell-border-left {{unless data.significant 'medium-gray'}}">
-        {{differenceSum}}
-      </td>
-      <td class="{{unless data.percent_significant 'medium-gray'}}">
-        {{differencePercent}}
-      </td>
-
-    {{/if}}
-  {{/if}}
-
-{{else}} {{! change over time mode }}
-
-  {{#if rowconfig.divider}}
-
-    <td>&nbsp;</td>
-    {{#unless comparison}}
-      <td class="cell-border-left">&nbsp;</td>
-      <td>&nbsp;</td>
-      <td class="cell-border-left">&nbsp;</td>
-      <td>&nbsp;</td>
-    {{/unless}}
-    <td class="cell-border-left">&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-    {{/if}}
-    <td>&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-    {{/if}}
-    {{#if comparison}}
-      <td class="cell-border-left">&nbsp;</td>
-      {{#if reliability}}
-        <td>&nbsp;</td>
-      {{/if}}
-      <td>&nbsp;</td>
-      {{#if reliability}}
-        <td>&nbsp;</td>
-      {{/if}}
-      <td class="cell-border-left">&nbsp;</td>
-      {{#if reliability}}
-        <td>&nbsp;</td>
-      {{/if}}
-      <td>&nbsp;</td>
-      {{#if reliability}}
-        <td>&nbsp;</td>
-      {{/if}}
-    {{/if}}
-
-  {{else}}{{! subheader && divider = false }}
-
-    <td>{{rowconfig.title}}</td>
-      {{#unless comparison}}
-        <td class="cell-border-left"><!--previous dataset sum-->
-        {{selectedEarlySum}}
-      </td>
-      <td><!--previous dataset percent-->
-        {{selectedEarlyPercent}}
-      </td>
-      <td class="cell-border-left"><!--current dataset sum-->
-        {{selectedCurrentSum}}
-      </td>
-      <td><!--current dataset percent-->
-      {{selectedCurrentPercent}}
-      </td>
-    {{/unless}}
-    <td class="cell-border-left"><!--selected area sum change-->
-      {{differenceChangeSum}}
-    </td>
-    {{#if reliability}}
-    <td><!--selected area sum change moe-->
-      {{differenceChangeSumMoE}}
-    </td>
-    {{/if}}
-    <td><!--selected area percent change-->
-      {{differenceChangePercent}}
-    </td>
-    {{#if reliability}}
-    <td><!--selected area percent change moe-->
-      {{differenceChangePercentMOE}}
-    </td>
-    {{/if}}
-    {{#if comparison}}
-    <td class="cell-border-left"><!--comparison area sum change-->
-      {{comparisonDifferenceChangeSum}}
-    </td>
-    {{#if reliability}}
-    <td><!--comparison area sum change moe-->
-        {{comparisonDifferenceChangeSumMoE}}
-    </td>
-    {{/if}}
-    <td><!--comparison area percent change-->
-        {{comparisonDifferenceChangePercent}}
-    </td><!--comparison area percent change moe-->
-    {{#if reliability}}
-      <td>
-          {{comparisonDifferenceChangePercentMOE}}
-      </td>
-      {{/if}}
-      <td class="cell-border-left">
-        {{longTermDifference}}
-      </td>
-      {{#if reliability}}
-        <td>
-          {{longTermDifferenceMOE}}
-        </td>
-      {{/if}}
-      <td>
-        {{longTermDifferencePercent}}
-      </td>
-      {{#if reliability}}
-      <td>
-        {{longTermDifferencePercentMOE}}
-      </td>
-      {{/if}}
-    {{/if}}
-  {{/if}}
-
-{{/if}}
-
-
-
-{{yield}}
