+++ app/templates/components/acs-table.hbs
@@ -1,24 +1,22 @@
 <div class="table-copy-ui show-for-large">
   <a class="button tiny gray text-orange" {{action 'handleCopy'}}>{{fa-icon 'clipboard'}} Copy Table to Clipboard</a>
 </div>
-
-{{#if media.isLarge}}
-  {{#sticky-element top=175 bottom=0}}
-    {{acs-table-header
+{{log mode}}
+{{#sticky-element top=stickyTop bottom=0}}
+  {{#if (eq mode 'current')}}
+    {{acs-table-header-current
       reliability=reliability
       comparison=comparison
       mode=mode
     }}
-  {{/sticky-element}}
-{{else}}
-  {{#sticky-element top=0 bottom=0}}
-    {{acs-table-header
+  {{else}}
+    {{acs-table-header-change
       reliability=reliability
       comparison=comparison
       mode=mode
     }}
-  {{/sticky-element}}
-{{/if}}
+  {{/if}}
+{{/sticky-element}}
 
 <div class="table-scroll">
   <table class="data-table">
@@ -27,25 +25,49 @@
         {{#if hasBlock}}
           {{yield}}
         {{else}}
-          {{acs-table-row
-            mode=mode
-            reliability=reliability
-            comparison=comparison
-            rowconfig=row
-            options=profile
-            data=
-              (unless 
-                row.special 
-                (get this.t1 row.data) 
-                (aggregate-special-variable rowConfig=row data=t1)
-              )
-            data2=
-              (unless 
-                row.special 
-                (get this.t2 row.data) 
-                (aggregate-special-variable rowConfig=row data=t2)
-              )
-          }}
+          {{#if (eq mode 'current')}}
+            {{acs-table-row-current
+              mode=mode
+              reliability=reliability
+              comparison=comparison
+              rowconfig=row
+              options=profile
+              data=
+                (unless
+                  row.special
+                  (get this.t1 row.data)
+                  (aggregate-special-variable rowConfig=row data=t1)
+                )
+              data2=
+                (unless
+                  row.special
+                  (get this.t2 row.data)
+                  (aggregate-special-variable rowConfig=row data=t2)
+                )
+            }}
+          {{else}}
+            {{acs-table-row-change
+              mode=mode
+              reliability=reliability
+              comparison=comparison
+              rowconfig=row
+              options=profile
+              data=
+                (unless
+                  row.special
+                  (get this.t1 row.data)
+                  (aggregate-special-variable rowConfig=row data=t1)
+                )
+              data2=
+                (unless
+                  row.special
+                  (get this.t2 row.data)
+                  (aggregate-special-variable rowConfig=row data=t2)
+                )
+            }}
+          {{/if}}
+
+
         {{/if}}
       {{/each}}
     </tbody>
